Fix EEPROM address variable types

+++ Multiprotocol/AFHDS2A_a7105.ino
@@ -298,7 +298,7 @@ uint16_t ReadAFHDS2A()
 				A7105_ReadData(AFHDS2A_RXPACKET_SIZE);
 				if(packet[0] == 0xbc && packet[9] == 0x01)
 				{
-					uint8_t addr;
+					uint16_t addr;
 					if(RX_num<16)
 						addr=AFHDS2A_EEPROM_OFFSET+RX_num*4;
 					else
@@ -433,7 +433,7 @@ uint16_t initAFHDS2A()
 	{
 		phase = AFHDS2A_DATA_INIT;
 		//Read RX ID from EEPROM based on RX_num, RX_num must be uniq for each RX
-		uint8_t addr;
+		uint16_t addr;
 		if(RX_num<16)
 			addr=AFHDS2A_EEPROM_OFFSET+RX_num*4;
 		else
