+++ Multiprotocol/HOTT_cc2500.ino
@@ -159,7 +159,7 @@ static void __attribute__((unused)) HOTT_init()
 	else
 	{
 		memcpy(&packet[40],rx_tx_addr,5);
-		uint8_t addr=HOTT_EEPROM_OFFSET+RX_num*5;
+		uint16_t addr=HOTT_EEPROM_OFFSET+RX_num*5;
 		debug("RXID: ");
 		for(uint8_t i=0;i<5;i++)
 		{
@@ -380,7 +380,7 @@ uint16_t ReadHOTT()
 						for(uint8_t i=0;i<HOTT_RX_PACKET_LEN;i++)
 							debug(" %02X", packet_in[i]);
 						debugln("");
-						uint8_t addr=HOTT_EEPROM_OFFSET+RX_num*5;
+						uint16_t addr=HOTT_EEPROM_OFFSET+RX_num*5;
 						for(uint8_t i=0; i<5; i++)
 							eeprom_write_byte((EE_ADDR)(addr+i),packet_in[5+i]);
 						BIND_DONE;
